+++ tests/testthat/test-graphs.R
@@ -96,7 +96,7 @@ test_that("path_graphs() works as expected", {
   graphs_lamp_prefs <- load_eeu_data() %>% 
     rebound_analysis() %>% 
     path_graphs(cases = "Lamp", 
-                   graph_types = ReboundTools::graph_types$preferences)
+                graph_types = ReboundTools::graph_types$preferences)
   expect_true(!is.null(graphs_lamp_prefs))
   
   # Try a preferences graph for lamps with fewer indifference curve points
@@ -115,6 +115,18 @@ test_that("path_graphs() works with grid_types = NULL", {
   graphs <- load_eeu_data() %>% 
     rebound_analysis() %>% 
     path_graphs(grid_types = NULL)
+  expect_true(!is.null(graphs))
+})
+
+
+test_that("path_graphs() works with show_indifference_curves = FALSE", {
+  pgp <- ReboundTools::path_graph_params
+  pgp$show_indifference_curves <- FALSE
+  graphs <- load_eeu_data() %>% 
+    dplyr::filter(.data[[ReboundTools::eeu_base_params$case]] == "Lamp") %>% 
+    rebound_analysis() %>% 
+    path_graphs(graph_types = ReboundTools::graph_types$preferences, 
+                graph_params = pgp)
   expect_true(!is.null(graphs))
 })