+++ NEWS.md
@@ -1,3 +1,11 @@
+# ReboundTools 0.1.26 (2021-04-23)
+
+* Added item `show_indifference_curves` to `graph_params` item of `path_graphs()` function.
+* New tests for the new argument.
+    - Now at 470 tests, all passing.
+    - Test coverage remains at 100%.
+
+
 # ReboundTools 0.1.26 (2021-04-22)
 
 * Fixed a bug where missing meta columns would cause graphing failure.